# bobCode/openapi/skills_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telecom Outage Resolution BOB",
    "description": "7-agent agentic AI system for telecom outage resolution. IBM BOB Hackathon 2026 POC.",
    "version": "0.1.0"
  },
  "paths": {
    "/orchestrate": {
      "post": {
        "tags": [
          "orchestration"
        ],
        "summary": "Run full 7-agent pipeline",
        "description": "Accepts a free-text customer complaint and runs it through the full 7-agent pipeline: Intent \u2192 Ticket \u2192 RCA \u2192 Escalation+Parallel \u2192 Resolution \u2192 Feedback. Returns aggregated result.",
        "operationId": "orchestrate_orchestrate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrchestrateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrchestrationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/webhook/orchestrate": {
      "post": {
        "tags": [
          "watson-orchestrate"
        ],
        "summary": "Watson Orchestrate skill entry point",
        "description": "Called by Watson Orchestrate when the Telecom BOB skill is invoked. Runs the full 7-agent pipeline and writes the resulting ticket to Cloudant via IBM BOB.",
        "operationId": "watson_orchestrate_webhook_webhook_orchestrate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatsonSkillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatsonSkillResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/webhook/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "watson-orchestrate",
          "watson-orchestrate"
        ],
        "summary": "Retrieve a BOB-created ticket from Cloudant",
        "description": "Return a ticket document from Cloudant by ticket_id.",
        "operationId": "get_ticket_webhook_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Ticket Webhook Tickets  Ticket Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrchestrateRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session identifier for this conversation turn"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id",
            "description": "Anonymised customer identifier"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "description": "Free-text customer complaint"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "customer_id",
          "message"
        ],
        "title": "OrchestrateRequest",
        "description": "Top-level request to the master orchestration endpoint."
      },
      "OrchestrationResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "ticket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id"
          },
          "intent_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Summary"
          },
          "ticket_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Summary"
          },
          "rca_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rca Summary"
          },
          "escalation_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Summary"
          },
          "analysis_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Summary"
          },
          "resolution_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Summary"
          },
          "feedback_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Summary"
          },
          "total_execution_ms": {
            "type": "number",
            "title": "Total Execution Ms",
            "default": 0.0
          },
          "agents_completed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Agents Completed"
          },
          "agents_failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Agents Failed"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "OrchestrationResult",
        "description": "Aggregated result returned after the full 7-agent pipeline completes."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WatsonSkillRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Watson Orchestrate session identifier"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id",
            "description": "Customer identifier from Watson Orchestrate context",
            "default": "WO-CUSTOMER"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "description": "Customer complaint forwarded from Watson Orchestrate"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Origination system identifier",
            "default": "watson_orchestrate"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "WatsonSkillRequest",
        "description": "Payload received from Watson Orchestrate when a skill is invoked.\nMaps Watson Orchestrate skill input parameters to our OrchestrateRequest."
      },
      "WatsonSkillResponse": {
        "properties": {
          "ticket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "queue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "root_cause_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause Summary"
          },
          "resolution_steps_count": {
            "type": "integer",
            "title": "Resolution Steps Count",
            "default": 0
          },
          "customer_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Message"
          },
          "escalated": {
            "type": "boolean",
            "title": "Escalated",
            "default": false
          },
          "sla_met": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Met"
          },
          "total_execution_ms": {
            "type": "number",
            "title": "Total Execution Ms",
            "default": 0.0
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          }
        },
        "type": "object",
        "title": "WatsonSkillResponse",
        "description": "Response returned to Watson Orchestrate after skill execution."
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}